22-put code into components
