+++ src/client/package.json
@@ -1,12 +1,8 @@
 {
-  "name": "client",
+  "name": "react-node-bagchal",
   "version": "0.1.0",
   "private": true,
-  "proxy": {
-    "/*": {
-      "target": "http://localhost:8080"
-    }
-  },
+  "proxy": "http://localhost:3001",
   "scripts": {
     "dev": "next dev --turbopack",
     "build": "next build",
@@ -14,23 +10,19 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "bootstrap": "^5.3.5",
-    "next": "^15.2.4",
+    "next": "^15.1.6",
     "react": "^19.0.0",
-    "react-bootstrap": "^2.10.9",
-    "react-dom": "^19.0.0",
-    "react-modal": "^3.16.3"
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
-    "@types/node": "20.17.30",
+    "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "@types/react-modal": "^3.16.3",
     "eslint": "^9",
     "eslint-config-next": "15.1.6",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
-    "typescript": "5.8.3"
+    "typescript": "^5"
   }
 }
